Refactor channel adapters into adapter module
RT88-79 local checkpoint.

Move Mastra channel adapter mechanics out of src/agents into src/adapters/channels. Add a generic initChannels() factory plus dedicated Linear, Slack, and GitHub connector modules. Keep supervisor-agent as the only channel attachment point for Palmer Linear while leaving orchestrator-agent disconnected from Linear.

Move webhook route helper exports to the adapter module and update Mastra server startup to import channelWebhookApiRoutesForAgents from the new boundary. Remove the old agent-owned channels.ts implementation.

Update channel tests to bundle src/adapters/channels/index.ts, preserve existing route/status coverage, add independent Slack/GitHub connector gating coverage, and assert orchestrator remains without channel routes.

Update Palmer/Linear docs to describe the adapter boundary, supervisor-only route ownership, and the lean Agent.channels -> Chat SDK adapter -> Postgres state architecture.

Verification performed before commit: npm test passed; npm run build passed; Mastra dev restarted in tmux session mastra-rt88-79; local unsigned supervisor Linear webhook probe returned 400 Missing webhook signature; orchestrator Linear webhook returned 404; Palmer replied to the RT88-79 smoke comment after restart.

Follow-up stability note: fresh-machine deployment remains at risk until the project has a preflight/doctor flow for DATABASE_URL reachability, model provider env, channel route registration, ACP generated runtime, webhook upstream URL, and the observed .mastra/output startup dependency issue.
