+++ mastra-agents/src/adapters/channels/github/create-github-channel.ts
@@ -1,0 +1,5 @@
+import { createGitHubAdapter } from "@chat-adapter/github";
+
+export function createGitHubChannel() {
+  return createGitHubAdapter();
+}
