+++ mastra-agents/src/adapters/channels/github/index.ts
@@ -1,0 +1,1 @@
+export { createGitHubChannel } from "./create-github-channel.js";
