+++ mastra-agents/src/adapters/channels/index.ts
@@ -1,0 +1,18 @@
+export {
+  initChannels,
+  listEnabledChannelPlatforms,
+  resolveAgentChannelStatus,
+  resolveChannelStatus,
+} from "./init-channels.js";
+export {
+  channelWebhookApiRoutesForAgents,
+  expectedChannelWebhookRoutes,
+  listConfiguredChannelWebhookRoutes,
+} from "./routes.js";
+export type {
+  AgentChannelsConfig,
+  ChannelPlatform,
+  ChannelStatus,
+  ChannelStatusMap,
+} from "./types.js";
+export { channelPlatforms } from "./types.js";
