+++ mastra-agents/src/adapters/channels/linear/index.ts
@@ -1,0 +1,1 @@
+export { createLinearChannel, getLinearMode } from "./create-linear-channel.js";
