+++ mastra-agents/src/adapters/channels/slack/create-slack-channel.ts
@@ -1,0 +1,5 @@
+import { createSlackAdapter } from "@chat-adapter/slack";
+
+export function createSlackChannel() {
+  return createSlackAdapter();
+}
