+++ mastra-agents/src/adapters/channels/slack/index.ts
@@ -1,0 +1,1 @@
+export { createSlackChannel } from "./create-slack-channel.js";
